Move identity-provider to separate module
