+++ package.json
@@ -13,6 +13,7 @@
     "node": ">=8.0.0"
   },
   "dependencies": {
+    "orbit-db-identity-provider": "github:orbitdb/orbit-db-identity-provider",
     "p-map": "^1.1.1",
     "p-whilst": "^1.0.0"
   },
