+++ test/signed-log.spec.js
@@ -6,8 +6,9 @@ const IPFSRepo = require('ipfs-repo')
 const DatastoreLevel = require('datastore-level')
 const Log = require('../src/log')
 const Keystore = require('orbit-db-keystore')
-const IdentityProvider = require('../src/identity-provider')
-const AccessController = require('../src/acl')
+// const Identity = require('../src/identity')
+const IdentityProvider = require('orbit-db-identity-provider')
+const AccessController = require('../src/default-access-controller')
 const startIpfs = require('./utils/start-ipfs')
 
 const apis = [require('ipfs')]
